docstring fix 2

+++ napari/utils/colormaps/colormap.py
@@ -153,7 +153,7 @@ class LabelColormap(Colormap):
     ----------
     seed : float
     use_selection : bool
-    selection : float
+    selection : int
     """
 
     seed: float = 0.5
